Remove remaining commands and related migration code
In a bid to clean this up, I've deleted all the migration information
so that we can make a clean move to the CKAN database migration code.
Although this is essentially the same, it's much more up to date than
the bits we have in this extension.

Users of this plugin are recommended to remove the DOI migration info
from the migrate_version table in the database too.