Fix coveralls install
Coveralls needs to be installed in the system running the tests on travis,
not on the docker container as we run coveralls from the host not the ckan
container. This should work because the repo root is mounted as a volume by
docker-compose so when the coverage data is written out it should be available
on the host.