fix: add new functions to support CKAN 2.10
In CKAN 2.10, they have renamed the IPackageController hooks. This really should have been a major change as it breaks the public API but hey ho here we are. To accommodate this while also supporting CKAN 2.9, I have renamed the hook functions we were using to their new CKAN 2.10 variants and then added functions which just call those functions which have the old CKAN 2.9 names. This is ok in this extension as we're not using any interfaces which have a name clash with the old names.

The old names should be removed as soon as we stop supporting CKAN 2.9 cause it's ugly stuff.